sound_service/requirements.txt
# ReSpeaker Sound Detection Service - Requirements

# Core USB communication
pyusb==1.1.0

# Audio processing
pyaudio>=0.2.11
numpy>=1.21.0
scipy>=1.7.0

# REST API
flask>=2.0.0
flask-cors>=3.0.10

# Beautiful CLI
rich>=13.0.0

# Utilities
python-dateutil>=2.8.0
